backend/requirements.txt
Django==5.2.7
django-ninja==1.3.0
SQLAlchemy==2.0.37
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0
PyMySQL==1.1.1
PyJWT==2.10.1
httpx==0.28.1
requests==2.32.3
beautifulsoup4==4.12.3
tldextract==5.1.3
python-whois==0.9.5
playwright==1.51.0
pytest==8.3.4
pytest-django==4.10.0
